feat: add more sections in home page and add some animations and effects
Added more sections in home page and some animations and effects in some components, like the CTA Button, hero title and in a section card. Need to replace the fake image to a 3D model, probably Ill use the spline to do that. After that, only need to refact some components to be responsive for all the screen sizes